multiple refactorings
