+++ .gitignore
@@ -8,3 +8,11 @@ gradle
 .gradle
 gradlew.bat
 gradlew
+bin/
+.vscode/
+src/main/resources/client-ssl.properties
+src/main/resources/ca.pem
+src/main/resources/client.keystore.p12
+src/main/resources/client.truststore.jks
+src/main/resources/service.cert
+src/main/resources/service.key
